RESTApi/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-appconf==1.0.4
django-cors-headers==3.5.0
django-ipware==3.0.1
django-restframework==0.0.1
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
cryptography~=3.1
Pillow~=7.2.0
PyPDF2~=1.26.0
six~=1.15.0
requests~=2.24.0